feat(columnar): configure column block size
Add (NumericPipelineSelector, int blockSize) constructor; validates power-of-2
>= 128. Threads blockSize through ColumNARDocValuesConsumer; NumericColumnWriter
derives it from NumericPipeline.blockSize() rather than a separate argument.

Introduce NumericPipelineTemplate so the selector returns a factory, keeping
blockSize ownership in the format.

Split server storage tests into a shared base, a pinned-byte-count suite, and
a diagnostic sweep. Parity ceilings asserted via MAX_COLUMNAR_TO_ES95_RATIO
(Ceiling record with separate bs128/bs512 thresholds). Randomize blockSize in
four unit test classes.

Add blockSize @Param to three JMH benchmarks. NumericFormat.codec() and
buildSegment() take blockSize; ES95 uses a workload-aware FieldContextResolver
matching ColumnarNumericStorageTestBase for apples-to-apples comparison.

+++ libs/columnar/src/main/java/org/elasticsearch/columnar/numeric/NumericColumnMetadata.java
@@ -229,6 +229,6 @@ public record NumericColumnMetadata(
     }
 
     int numBlocks() {
-        return numValues == 0 ? 0 : (numValues + blockSize - 1) / blockSize;
+        return numValues == 0 ? 0 : (int) ((numValues + (long) blockSize - 1) / blockSize);
     }
 }
